No issue: RTL nav bar UI tests
